refactor(sequencer): extract initialization into builder pattern
Extract PreferredSequencer initialization logic into a dedicated
PreferredSequencerBuilder to improve code organization and maintainability.

Changes:
- Add preferred/initialization.rs with PreferredSequencerBuilder
- PreferredSequencer::create() now delegates to builder
- Reduce mod.rs from 1,498 to 1,245 LOC (17% reduction)
- Public API remains unchanged (backward compatible)

This is the first step in a larger refactoring plan to improve
the sequencer codebase organization and testability.
